@@ -1,12 +1,16 @@
 #include <time.h>
+
 #include <condition_variable>
 #include <cstdlib>
 #include <iostream>
 #include <mutex>
 #include <thread>
-// #include "TimeMeter.hh"
+//#include "TimeMeter.hh"
+#include <memory>
+
 #include "Philo.hpp"
 #include "fork.hpp"
+
 using namespace std;
 
 std::mutex mutex_var;
@@ -16,14 +20,11 @@ constexpr unsigned int N_PHILOSOPHES = 5;
 auto LEFT_FORK = [](unsigned int N) { return N; };
 auto RIGHT_FORK = [](unsigned int N) { return ((N + 1) % N_PHILOSOPHES); };
 
-Philosopher* Philosopher[N_PHILOSOPHES]{nullptr};
-Fork* forchetta[N_PHILOSOPHES]{nullptr};
-
 auto print_status = [](const unsigned int& n, const std::string& message) {
   std::cout << "Philosopher " << n << " " << message << "\n";
 };
 
-void philosopher(unsigned int n) {
+void philosopher(unsigned int n, Philosopher* Philosopher, Fork* forchetta) {
   {
     // barrier for wait a go signal
     std::unique_lock<std::mutex> lock(mutex_var);
@@ -34,57 +35,59 @@ void philosopher(unsigned int n) {
 
   while (true) {  // infinity loop
     // thinking
-    Philosopher[n]->thinking();
+    Philosopher[n].thinking();
     print_status(n, "is THINKING");
     std::chrono::milliseconds duration(500);
     std::this_thread::sleep_for(duration);
 
     // enter in the critical section
-
-    // possible deadlock
-    Philosopher[n]->SetLeftFork(forchetta[LEFT_FORK(n)]);
-    print_status(n, "acquired the left fork!");
     mutex_var.lock();
-    if (forchetta[RIGHT_FORK(n)]->getStateFork() == forkstate::FREE) {
+    // possible deadlock
+    Philosopher[n].SetLeftFork(&forchetta[LEFT_FORK(n)]);
+    print_status(n, "acquired the left fork!");
+    if (forchetta[LEFT_FORK(n)].getStateFork() == forkstate::INUSE) {
+      Philosopher[n].SetRightFork(&forchetta[RIGHT_FORK(n)]);
       print_status(n, "acquired the right fork!");
-      if (!Philosopher[n]->getCanEat()){
-          print_status(n, " waiting for the right fork!\n");
-          Philosopher[n]->ReleaseLeftFork(forchetta[RIGHT_FORK(n)]);
-          print_status(n, " then release fork!\n");
+      Philosopher[n].setCanEat(&forchetta[LEFT_FORK(n)],
+                               &forchetta[RIGHT_FORK(n)]);
+      if (Philosopher[n].getCanEat() == false) {
+        print_status(n, " waiting for the right fork!\n");
+        Philosopher[n].ReleaseLeftFork(&forchetta[LEFT_FORK(n)]);
+        Philosopher[n].ReleaseRightFork(&forchetta[RIGHT_FORK(n)]);
+        print_status(n, " then release fork!\n");
       }
     }
-    mutex_var.unlock();
 
     // eating
-    if (Philosopher[n]->getCanEat()) {
-        mutex_var.lock();
+    if (Philosopher[n].getCanEat()) {
+      Philosopher[n].eating();
       print_status(n, "is EATING");
       std::chrono::milliseconds duration1(1000);
       std::this_thread::sleep_for(duration1);
       // enter in the critical section
-      Philosopher[n]->ReleaseLeftFork(forchetta[LEFT_FORK(n)]);
+      Philosopher[n].ReleaseLeftFork(&forchetta[LEFT_FORK(n)]);
       print_status(n, "drop the right fork!");
-      Philosopher[n]->ReleaseRightFork(forchetta[RIGHT_FORK(n)]);
+      Philosopher[n].ReleaseRightFork(&forchetta[RIGHT_FORK(n)]);
       print_status(n, "drop the left fork!");
-      mutex_var.unlock();  // exit from the critical section
     }
+    mutex_var.unlock();  // exit from the critical section
 
     // sleeping
-    Philosopher[n]->sleeping();
+    Philosopher[n].sleeping();
     print_status(n, "is SLEEPING\n");
     std::chrono::milliseconds duration2(1500);
     std::this_thread::sleep_for(duration2);
   }
 }
 
-void monitor() {
+void monitor(Philosopher* Philosopher, Fork* forchetta) {
   while (true) { /* ciclo infinito */
     std::chrono::milliseconds duration(1000);
     std::this_thread::sleep_for(duration);
     // enter in the critical section
     cout << "\nPhilosophes:  ";
     for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-      switch (Philosopher[i]->GetPhilosopherState()) {
+      switch (Philosopher[i].GetPhilosopherState()) {
         case action::THINK:
           cout << " T";
           break;  // think
@@ -95,12 +98,12 @@ void monitor() {
           cout << " S";
           break;  // sleep
         default:
-          break;
+          throw "Error";
       }
     }
     std::cout << "\nForks:        ";
     for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-      switch (forchetta[i]->getStateFork()) {
+      switch (forchetta[i].getStateFork()) {
         case forkstate::FREE:
           cout << " F";
           break;
@@ -108,47 +111,52 @@ void monitor() {
           cout << " U";
           break;
         default:
-          break;
+          throw "Error";
       }
     }
 
     // Print the number philosophes eat
-     std::cout << "\nTime Eat:      ";
+    std::cout << "\nTime Eat:      ";
     for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-      cout << Philosopher[i]->getEat() << " ";
+      std::cout << Philosopher[i].getEat() << " ";
     }
     // Print the number philosophes sleep
     std::cout << endl;
     std::cout << "Time Sleep:    ";
     for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-      std::cout << Philosopher[i]->getSleep() << " ";
+      std::cout << Philosopher[i].getSleep() << " ";
     }
     // Print the number philosophes think
     std::cout << std::endl;
     std::cout << "Time Think:    ";
     for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-      std::cout << Philosopher[i]->getThink() << " ";
+      std::cout << Philosopher[i].getThink() << " ";
     }
     std::cout << std::endl << std::endl;
   }
 }
 
 int main() {
-  for (unsigned int i = 0; i < N_PHILOSOPHES; i++) {
-    forchetta[i] = new class Fork();
-    Philosopher[i] = new class Philosopher();
-  }
+  //  Philosopher* Philosopher[N_PHILOSOPHES]{nullptr};
+  //  Fork* forchetta[N_PHILOSOPHES]{nullptr};
+  auto phil = std::make_unique<Philosopher[]>(N_PHILOSOPHES);
+  auto forchetta = std::make_unique<Fork[]>(N_PHILOSOPHES);
+
+  //  for (unsigned int i = 0; i < N_PHILOSOPHES; i++) {
+  //    forchetta[i] = new class Fork();
+  //    Philosopher[i] = new class Philosopher();
+  //  }
 
   // create threads for philosophes
   for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
     cout << "Create the philosopher " << i << "\n";
-    std::thread(philosopher, i).detach();
+    std::thread(philosopher, i, phil.get(), forchetta.get()).detach();
     std::chrono::milliseconds duration(1000);
     std::this_thread::sleep_for(duration);
   }
 
   // create threads for monitor
-  std::thread(monitor).detach();
+  std::thread(monitor, phil.get(), forchetta.get()).detach();
   std::chrono::seconds duration(10);
   std::this_thread::sleep_for(duration);
 
@@ -159,10 +167,10 @@ int main() {
   std::this_thread::sleep_for(duration1);
 
   // erase the dynamic memory
-  for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
-    delete forchetta[i];
-    delete Philosopher[i];
-  }
+  //  for (unsigned int i = 0; i < N_PHILOSOPHES; ++i) {
+  //    delete forchetta[i];
+  //    delete Philosopher[i];
+  //  }
 
   return 0;
 }
